requirements3.txt
numpy
pandas
matplotlib
yfinance
requests
python-dotenv
